backend/app/stagehand-action/requirements.txt
playwright
python-dotenv
stagehand-py
rich 